fix(kafka): Refactor the code
